Move util.contest and sampler into audits module
