+++ tests/test_sampler.py
@@ -1,7 +1,6 @@
 import pytest
-import sampler
-
-from util.contest import Contest as Contest
+from audits import sampler
+from audits.sampler_contest import Contest
 
 seed = "12345678901234567890abcdefghijklmnopqrstuvwxyz😊"
 risk_limit = 0.1
